refactoring pub/sub
